src/services/gateway/requirements.txt
httpx
fastapi-standalone-docs
fastapi[standard]
sqlalchemy
asyncpg
python-dotenv
pydantic-settings
pyjwt
passlib
bcrypt~=4.3.0
faststream[rabbit]
faststream[cli]
loguru
uvicorn
starlette
pydantic
pydantic_core
redis
pytest
pytest-asyncio
